+++ src/calendar-list/style.ts
@@ -1,6 +1,6 @@
 import {Platform, StyleSheet} from 'react-native';
 import * as defaultStyle from '../style';
-import {Theme} from '../commons/types';
+import {Theme} from '../types';
 
 export default function getStyle(theme: Theme = {}) {
   const appStyle = {...defaultStyle, ...theme};