+++ src/calendar/day/basic/index.tsx
@@ -3,14 +3,15 @@ import PropTypes from 'prop-types';
 
 import React, {Component, Fragment} from 'react';
 import {TouchableOpacity, Text, View} from 'react-native';
+
+import {Theme, DateData, DayState} from '../../../types';
 // @ts-expect-error
 import {shouldUpdate} from '../../../component-updater';
 import styleConstructor from './style';
 import Marking, {MarkingTypes, MarkingProps} from '../marking';
-import {Theme} from '../../../commons/types';
 
 export interface BasicDayProps {
-  state?: 'selected' | 'disabled' | 'today';
+  state?: DayState;
   /** The marking object */
   marking?: MarkingProps;
   /** Date marking style [simple/period/multi-dot/multi-period]. Default = 'simple' */
@@ -18,13 +19,15 @@ export interface BasicDayProps {
   /** Theme object */
   theme?: Theme;
   /** onPress callback */
-  onPress?: (date: Date) => void;
+  onPress?: (date: DateData) => void;
   /** onLongPress callback */
   onLongPress?: (date: Date) => void;
   /** The date to return from press callbacks */
   date?: Date;
   /** Disable all touch events for disabled days. can be override with disableTouchEvent in markedDates*/
   disableAllTouchEventsForDisabledDays?: boolean;
+  /** Disable all touch events for inactive days. can be override with disableTouchEvent in markedDates*/
+  disableAllTouchEventsForInactiveDays?: boolean;
   /** Test ID*/
   testID?: string;
   /** Accessibility label */
@@ -35,7 +38,7 @@ export default class BasicDay extends Component<BasicDayProps> {
   static displayName = 'IGNORE';
 
   static propTypes = {
-    state: PropTypes.oneOf(['selected', 'disabled', 'today', '']),
+    state: PropTypes.oneOf(['selected', 'disabled', 'inactive', 'today', '']),
     /** The marking object */
     marking: PropTypes.any,
     /** Date marking style [simple/period/multi-dot/multi-period]. Default = 'simple' */
@@ -49,7 +52,10 @@ export default class BasicDay extends Component<BasicDayProps> {
     /** The date to return from press callbacks */
     date: PropTypes.object,
     /** Disable all touch events for disabled days. Can be override with disableTouchEvent in markedDates*/
-    disableAllTouchEventsForDisabledDays: PropTypes.bool
+    disableAllTouchEventsForDisabledDays: PropTypes.bool,
+    /** Disable all touch events for inactive days. can be override with disableTouchEvent in markedDates*/
+    disableAllTouchEventsForInactiveDays: PropTypes.bool
+
   };
 
   style = styleConstructor(this.props.theme);
@@ -79,7 +85,7 @@ export default class BasicDay extends Component<BasicDayProps> {
   }
 
   shouldDisableTouchEvent() {
-    const {disableAllTouchEventsForDisabledDays} = this.props;
+    const {disableAllTouchEventsForDisabledDays, disableAllTouchEventsForInactiveDays} = this.props;
     const {disableTouchEvent} = this.marking;
     let disableTouch = false;
 
@@ -87,6 +93,8 @@ export default class BasicDay extends Component<BasicDayProps> {
       disableTouch = disableTouchEvent;
     } else if (typeof disableAllTouchEventsForDisabledDays === 'boolean' && this.isDisabled()) {
       disableTouch = disableAllTouchEventsForDisabledDays;
+    } else if (typeof disableAllTouchEventsForInactiveDays === 'boolean' && this.isInactive()) {
+      disableTouch = disableAllTouchEventsForInactiveDays;
     }
     return disableTouch;
   }
@@ -97,6 +105,10 @@ export default class BasicDay extends Component<BasicDayProps> {
 
   isDisabled() {
     return typeof this.marking.disabled !== 'undefined' ? this.marking.disabled : this.props.state === 'disabled';
+  }
+
+  isInactive() {
+    return this.marking?.inactive;
   }
 
   isToday() {
@@ -152,6 +164,8 @@ export default class BasicDay extends Component<BasicDayProps> {
       style.push(this.style.disabledText);
     } else if (this.isToday()) {
       style.push(this.style.todayText);
+    } else if (this.isInactive()) {
+      style.push(this.style.inactiveText);
     }
 
     //Custom marking type
@@ -173,6 +187,7 @@ export default class BasicDay extends Component<BasicDayProps> {
         marked={this.isMultiDot() ? true : marked}
         selected={this.isSelected()}
         disabled={this.isDisabled()}
+        inactive={this.isInactive()}
         today={this.isToday()}
         dotColor={dotColor}
         dots={dots}