+++ src/calendar/header/style.ts
@@ -1,6 +1,6 @@
 import {StyleSheet, Platform} from 'react-native';
 import * as defaultStyle from '../../style';
-import {Theme} from '../../commons/types';
+import {Theme} from '../../types';
 
 export default function (theme: Theme = {}) {
   const appStyle = {...defaultStyle, ...theme};