+++ src/calendar/index.tsx
@@ -4,12 +4,12 @@ import XDate from 'xdate';
 import memoize from 'memoize-one';
 
 import React, {Component, RefObject} from 'react';
-import {View, ViewStyle} from 'react-native';
+import {View, ViewStyle, StyleProp} from 'react-native';
 // @ts-expect-error
 import GestureRecognizer, {swipeDirections} from 'react-native-swipe-gestures';
 
 // @ts-expect-error
-import dateutils from '../dateutils';
+import {page, isGTE, isLTE, sameMonth} from '../dateutils';
 // @ts-expect-error
 import {xdateToData, parseDate, toMarkingFormat} from '../interface';
 // @ts-expect-error
@@ -19,23 +19,22 @@ import {getState} from '../day-state-manager';
 import {extractComponentProps} from '../component-updater';
 // @ts-expect-error
 import {WEEK_NUMBER} from '../testIDs';
+import {Theme, DateData} from '../types';
 import styleConstructor from './style';
 import CalendarHeader, {CalendarHeaderProps} from './header';
 import Day, {DayProps} from './day/index';
 import BasicDay from './day/basic';
 import {MarkingProps} from './day/marking';
-import {Theme} from '../commons/types';
-
 
 type MarkedDatesType = {
-  [key: string]: MarkingProps
-}
+  [key: string]: MarkingProps;
+};
 
 export interface CalendarProps extends CalendarHeaderProps, DayProps {
   /** Specify theme properties to override specific styles for calendar parts */
   theme?: Theme;
   /** Specify style for calendar container element */
-  style?: ViewStyle;
+  style?: StyleProp<ViewStyle>;
   /** Initially visible month */
   current?: XDate;
   /** Minimum date that can be selected, dates before minDate will be grayed out */
@@ -55,13 +54,13 @@ export interface CalendarProps extends CalendarHeaderProps, DayProps {
   /** Always show six weeks on each month (only when hideExtraDays = false) */
   showSixWeeks?: boolean;
   /** Handler which gets executed on day press */
-  onDayPress?: (date: Date) => any;
+  onDayPress?: (date: DateData) => void;
   /** Handler which gets executed on day long press */
-  onDayLongPress?: (date: Date) => any;
+  onDayLongPress?: (date: DateData) => void;
   /** Handler which gets executed when month changes in calendar */
-  onMonthChange?: () => void;
+  onMonthChange?: () => DateData;
   /** Handler which gets executed when visible month changes in calendar */
-  onVisibleMonthsChange?: () => void;
+  onVisibleMonthsChange?: (months: DateData[]) => void;
   /** Disables changing month when click on days of other months (when hideExtraDays is false) */
   disableMonthChange?: boolean;
   /** Enable the option to swipe between months */
@@ -75,7 +74,7 @@ export interface CalendarProps extends CalendarHeaderProps, DayProps {
 }
 
 interface CalendarState {
-  currentMonth: any
+  currentMonth: any;
 }
 /**
  * @description: Calendar component
@@ -157,13 +156,13 @@ class Calendar extends Component<CalendarProps, CalendarState> {
     });
   };
 
-  handleDayInteraction(date: Date, interaction?: (date: Date) => any) {
+  handleDayInteraction(date: Date, interaction?: (date: DateData) => void) {
     const {disableMonthChange} = this.props;
     const day = parseDate(date);
     const minDate = parseDate(this.props.minDate);
     const maxDate = parseDate(this.props.maxDate);
 
-    if (!(minDate && !dateutils.isGTE(day, minDate)) && !(maxDate && !dateutils.isLTE(day, maxDate))) {
+    if (!(minDate && !isGTE(day, minDate)) && !(maxDate && !isLTE(day, maxDate))) {
       const shouldUpdateMonth = disableMonthChange === undefined || !disableMonthChange;
 
       if (shouldUpdateMonth) {
@@ -229,7 +228,7 @@ class Calendar extends Component<CalendarProps, CalendarState> {
     const {hideExtraDays, markedDates} = this.props;
     const dayProps = extractComponentProps(Day, this.props);
 
-    if (!dateutils.sameMonth(day, this.state.currentMonth) && hideExtraDays) {
+    if (!sameMonth(day, this.state.currentMonth) && hideExtraDays) {
       return <View key={id} style={this.style.emptyDayContainer} />;
     }
 
@@ -269,7 +268,7 @@ class Calendar extends Component<CalendarProps, CalendarState> {
     const {currentMonth} = this.state;
     const {firstDay, showSixWeeks, hideExtraDays} = this.props;
     const shouldShowSixWeeks = showSixWeeks && !hideExtraDays;
-    const days = dateutils.page(currentMonth, firstDay, shouldShowSixWeeks);
+    const days = page(currentMonth, firstDay, shouldShowSixWeeks);
     const weeks = [];
 
     while (days.length) {
@@ -286,7 +285,7 @@ class Calendar extends Component<CalendarProps, CalendarState> {
 
     if (current) {
       const lastMonthOfDay = toMarkingFormat(current.clone().addMonths(1, true).setDate(1).addDays(-1));
-      if (displayLoadingIndicator && !(markedDates?.[lastMonthOfDay])) {
+      if (displayLoadingIndicator && !markedDates?.[lastMonthOfDay]) {
         indicator = true;
       }
     }