+++ src/expandableCalendar/agendaList.tsx
@@ -1,15 +1,42 @@
 import _ from 'lodash';
-import React, {Component} from 'react';
-import {SectionList, Text} from 'react-native';
 import PropTypes from 'prop-types';
 import XDate from 'xdate';
-import dateutils from '../dateutils';
+
+import React, {Component} from 'react';
+import {Text, SectionList, SectionListProps, DefaultSectionT, SectionListData, ViewStyle, NativeSyntheticEvent, NativeScrollEvent, LayoutChangeEvent, ViewToken} from 'react-native';
+
+// @ts-expect-error
+import {isToday} from '../dateutils';
+// @ts-expect-error
+import {getMoment} from '../momentResolver';
+import {Theme} from '../types';
 import styleConstructor from './style';
 import asCalendarConsumer from './asCalendarConsumer';
-import {getMoment} from '../momentResolver';
+
 
 const commons = require('./commons');
-const UPDATE_SOURCES = commons.UPDATE_SOURCES;
+const updateSources = commons.UpdateSources;
+
+interface Props extends SectionListProps<any, DefaultSectionT> {
+  /** day format in section title. Formatting values: http://arshaw.com/xdate/#Formatting */
+  dayFormat?: string;
+  /** a function to custom format the section header's title */
+  dayFormatter?: (arg0: string) => string;
+  /** whether to use moment.js for date string formatting
+   * (remember to pass 'dayFormat' with appropriate format, like 'dddd, MMM D') */
+  useMoment?: boolean;
+  /** whether to mark today's title with the "Today, ..." string. Default = true */
+  markToday?: boolean;
+  /** style passed to the section view */
+  sectionStyle?: ViewStyle;
+  /** whether to block the date change in calendar (and calendar context provider) when agenda scrolls */
+  avoidDateUpdates?: boolean;
+  /** offset scroll to section */
+  viewOffset?: number;
+  theme?: Theme;
+  context?: any;
+}
+export type AgendaListProps = Props;
 
 /**
  * @description: AgendaList component
@@ -17,11 +44,11 @@ const UPDATE_SOURCES = commons.UPDATE_SOURCES;
  * @extends: SectionList
  * @example: https://github.com/wix/react-native-calendars/blob/master/example/src/screens/expandableCalendar.js
  */
-class AgendaList extends Component {
+class AgendaList extends Component<Props> {
   static displayName = 'AgendaList';
 
   static propTypes = {
-    ...SectionList.propTypes,
+    // ...SectionList.propTypes,
     /** day format in section title. Formatting values: http://arshaw.com/xdate/#Formatting */
     dayFormat: PropTypes.string,
     /** a function to custom format the section header's title */
@@ -43,21 +70,42 @@ class AgendaList extends Component {
     markToday: true
   };
 
-  constructor(props) {
-    super(props);
-    this.style = styleConstructor(props.theme);
+  style = styleConstructor(this.props.theme);
+  _topSection = _.get(this.props, 'sections[0].title');
+  didScroll = false;
+  sectionScroll = false;
+  viewabilityConfig = {
+    itemVisiblePercentThreshold: 20 // 50 means if 50% of the item is visible
+  };
+  list: React.RefObject<any> = React.createRef();
+  sectionHeight = 0;
 
-    this._topSection = _.get(props, 'sections[0].title');
-    this.didScroll = false;
-    this.sectionScroll = false;
-
-    this.viewabilityConfig = {
-      itemVisiblePercentThreshold: 20 // 50 means if 50% of the item is visible
-    };
-    this.list = React.createRef();
+  componentDidMount() {
+    const {date} = this.props.context;
+    if (date !== this._topSection) {
+      setTimeout(() => {
+        const sectionIndex = this.getSectionIndex(date);
+        if (sectionIndex) {
+          this.scrollToSection(sectionIndex);
+        }
+      }, 500);
+    }
   }
 
-  getSectionIndex(date) {
+  componentDidUpdate(prevProps: Props) {
+    const {updateSource, date} = this.props.context;
+    if (date !== prevProps.context.date) {
+      // NOTE: on first init data should set first section to the current date!!!
+      if (updateSource !== updateSources.LIST_DRAG && updateSource !== updateSources.CALENDAR_INIT) {
+        const sectionIndex = this.getSectionIndex(date);
+        if (sectionIndex) {
+          this.scrollToSection(sectionIndex);
+        }
+      }
+    }
+  }
+
+  getSectionIndex(date: Date) {
     let i;
     _.map(this.props.sections, (section, index) => {
       // NOTE: sections titles should match current date format!!!
@@ -69,28 +117,7 @@ class AgendaList extends Component {
     return i;
   }
 
-  componentDidMount() {
-    const {date} = this.props.context;
-    if (date !== this._topSection) {
-      setTimeout(() => {
-        const sectionIndex = this.getSectionIndex(date);
-        this.scrollToSection(sectionIndex);
-      }, 500);
-    }
-  }
-
-  componentDidUpdate(prevProps) {
-    const {updateSource, date} = this.props.context;
-    if (date !== prevProps.context.date) {
-      // NOTE: on first init data should set first section to the current date!!!
-      if (updateSource !== UPDATE_SOURCES.LIST_DRAG && updateSource !== UPDATE_SOURCES.CALENDAR_INIT) {
-        const sectionIndex = this.getSectionIndex(date);
-        this.scrollToSection(sectionIndex);
-      }
-    }
-  }
-
-  getSectionTitle(title) {
+  getSectionTitle(title: string) {
     if (!title) return;
 
     const {dayFormatter, dayFormat, useMoment, markToday} = this.props;
@@ -103,77 +130,88 @@ class AgendaList extends Component {
         const moment = getMoment();
         sectionTitle = moment(title).format(dayFormat);
       } else {
-        sectionTitle = XDate(title).toString(dayFormat);
+        sectionTitle = new XDate(title).toString(dayFormat);
       }
     }
 
     if (markToday) {
+      // @ts-expect-error
       const todayString = XDate.locales[XDate.defaultLocale].today || commons.todayString;
-      const isToday = dateutils.isToday(XDate(title));
-      sectionTitle = isToday ? `${todayString}, ${sectionTitle}` : sectionTitle;
+      const today = isToday(new XDate(title));
+      sectionTitle = today ? `${todayString}, ${sectionTitle}` : sectionTitle;
     }
 
     return sectionTitle;
   }
 
-  scrollToSection(sectionIndex) {
-    if (this.list.current && sectionIndex !== undefined) {
+  scrollToSection(sectionIndex: number) {
+    if (this.list?.current && sectionIndex !== undefined) {
+      const {sections, viewOffset = 0} = this.props;
       this.sectionScroll = true; // to avoid setDate() in onViewableItemsChanged
-      this._topSection = this.props.sections[sectionIndex].title;
+      this._topSection = sections[sectionIndex].title;
 
       this.list.current.scrollToLocation({
         animated: true,
         sectionIndex: sectionIndex,
         itemIndex: 0,
         viewPosition: 0, // position at the top
-        viewOffset: (commons.isAndroid ? this.sectionHeight : 0) + this.props.viewOffset
+        viewOffset: (commons.isAndroid ? this.sectionHeight : 0) + viewOffset
       });
     }
   }
 
-  onViewableItemsChanged = ({viewableItems}) => {
-    if (viewableItems && !this.sectionScroll) {
-      const topSection = _.get(viewableItems[0], 'section.title');
+  onViewableItemsChanged = ((info: {viewableItems: Array<ViewToken>; changed: Array<ViewToken>}) => {
+    if (info?.viewableItems && !this.sectionScroll) {
+      const topSection = _.get(info?.viewableItems[0], 'section.title');
       if (topSection && topSection !== this._topSection) {
         this._topSection = topSection;
         if (this.didScroll && !this.props.avoidDateUpdates) {
           // to avoid setDate() on first load (while setting the initial context.date value)
-          _.invoke(this.props.context, 'setDate', this._topSection, UPDATE_SOURCES.LIST_DRAG);
+          _.invoke(this.props.context, 'setDate', this._topSection, updateSources.LIST_DRAG);
         }
       }
     }
-  };
+  });
 
-  onScroll = event => {
+  onScroll = (event: NativeSyntheticEvent<NativeScrollEvent>) => {
     if (!this.didScroll) {
       this.didScroll = true;
     }
     _.invoke(this.props, 'onScroll', event);
   };
 
-  onMomentumScrollBegin = event => {
+  onMomentumScrollBegin = (event: NativeSyntheticEvent<NativeScrollEvent>) => {
     _.invoke(this.props.context, 'setDisabled', true);
     _.invoke(this.props, 'onMomentumScrollBegin', event);
   };
 
-  onMomentumScrollEnd = event => {
+  onMomentumScrollEnd = (event: NativeSyntheticEvent<NativeScrollEvent>) => {
     // when list momentum ends AND when scrollToSection scroll ends
     this.sectionScroll = false;
     _.invoke(this.props.context, 'setDisabled', false);
     _.invoke(this.props, 'onMomentumScrollEnd', event);
   };
 
-  onScrollToIndexFailed = info => {
-    console.warn('onScrollToIndexFailed info: ', info);
+  onScrollToIndexFailed = (info: {
+    index: number;
+    highestMeasuredFrameIndex: number;
+    averageItemLength: number;
+}) => {
+    if (this.props.onScrollToIndexFailed) {
+      this.props.onScrollToIndexFailed(info);
+    } else {
+      console.warn('onScrollToIndexFailed info: ', info);
+    }  
   };
 
-  onHeaderLayout = ({nativeEvent}) => {
-    this.sectionHeight = nativeEvent.layout.height;
+  onHeaderLayout = (event: LayoutChangeEvent) => {
+    this.sectionHeight = event.nativeEvent.layout.height;
   };
 
-  renderSectionHeader = ({section: {title}}) => {
+  renderSectionHeader = (info: {section: SectionListData<any, DefaultSectionT>}) => {
     const {renderSectionHeader, sectionStyle} = this.props;
-
+    const title = info?.section?.title;
+    
     if (renderSectionHeader) {
       return renderSectionHeader(title);
     }
@@ -185,7 +223,7 @@ class AgendaList extends Component {
     );
   };
 
-  keyExtractor = (item, index) => {
+  keyExtractor = (item: any, index: number) => {
     const {keyExtractor} = this.props;
     return _.isFunction(keyExtractor) ? keyExtractor(item, index) : String(index);
   };