+++ CHANGELOG.md
@@ -702,3 +702,25 @@
 - TS - Fix build process (PR #1545).
 - ExpandableCalendar - fix arrow press callbacks (PR #1489).
 - README & ISSUE_TEMPLATE - edit.
+
+
+## [1.1265.0] - 2021-8-23
+
+## Changed
+- Change import to require - solved "Cannot assign to read only property 'exports' of object '#<Object>'" (PR #1553).
+- TS migration for all Calendars components.
+- Upgrading to react-native 0.64.2.
+
+## Added
+- AgendaList - support offset scroll to section (PR #1526).
+- AgendaList - allow passing 'onScrollToIndexFailed' (PR #1594).
+- Day - support 'inactive' state and add 'disableAllTouchEventsForInactiveDays' prop (PR #1596).
+
+## Fixed
+- ExpandableCalendar - 'onDayPress' not called (PR #1584).
+- Fixed issue with 'dateutils' export (PR #1548).
+
+## [1.1266.0] - 2021-8-24
+
+## Changed
+- Disabling Detox e2e tests.